fix: add search-kernel command
